Semaine5/01_RAG_GPT/requirements.txt
# coeur
gradio>=5.0
langchain>=0.2.12
langchain-openai>=0.1.25
langchain-community>=0.2.11
langchain-chroma>=0.2.5
openai>=1.40
tiktoken>=0.7

# vectordb
chromadb>=1.0.9

# pdf
pypdf>=4.2.0
pandas>=2.2

# util
python-dotenv>=1.0
